Add CommonJS support via dual ESM/CJS packaging
Enable CommonJS consumers to require tinyqueue by adding an index.cjs
entry point with the exports field for conditional resolution. The ESM
index.js re-exports from index.cjs to keep a single source of truth.
